refactor: extract file mapping utilities
